Flatten registered symbols
N.B.  This patch changes the contract between workflows and is thus a breaking change

The current code registers each defworkflow/defactivity with a fully qualified name, and this
becomes part of the Temporal contract.  However, Temporal already guarantees uniquness on a per
namespace+taskqueue basis.  Therefore, having the workflow/activity also be fully qualified
introduces an extra degree of fragility, because simply refactoring clojure namespaces changes
the contract.

Other SDKs like Golang use the workflow/activity function name (alone, without the package prefix)
as the contract, so this brings the Clojure SDK in line with prior practices and elimintes a source
of fragility at the expense of requiring the program to uniquely name symbols even across namespaces.

Signed-off-by: Greg Haskins <greg@manetu.com>