args serdes should adhere to this SDKs protocol
Args typically are serialized and deserialized using objarray-> and
args-> (defined in temporal.internal.utils) so that data-conversion can
transparently work with arbitrary data.  Schedule should adhere to this
protocol and apply objarray-> on arguments so as to satisfy the args->
call when workflow is executed.

TODO: enhance UT to actually launch a workflow when base java library
adds proper schedule support to its "testing" framework.

Signed-off-by: Srinivasan Muralidharan <smuralidharan@manetu.com>